+++ app/helpers/paxful_engine/application_helper.rb
@@ -6,5 +6,10 @@ module PaxfulEngine
       status == "successful" ? "text-success" : "text-danger"
     end
 
+    def format_datetime(datetime, timezone = "Asia/Manila")
+      return "-" if datetime.nil?
+      datetime.in_time_zone(timezone).strftime("%m/%d/%Y %I:%M%p")
+    end
+
   end
 end